IDEMPIERE-6154: dropped ProjectIssueUtils class
