+++ org.adempiere.base/src/org/compiere/model/MInOutLine.java
@@ -18,6 +18,7 @@ package org.compiere.model;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
+import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.util.List;
 import java.util.Properties;
@@ -959,4 +960,74 @@ public class MInOutLine extends X_M_InOutLine
 		
 		return success;
 	}
+
+	/**
+	 * Check Storage On Hand Qty and Return
+	 *
+	 * @param  inOutLine
+	 * @param  requiredQty
+	 * @param  asiID
+	 * @return
+	 */
+	public static BigDecimal getInOutLineQTY(MInOutLine inOutLine, BigDecimal requiredQty, int asiID)
+	{
+		MStorageOnHand storageOnHand = MStorageOnHand.get(	inOutLine.getCtx(), inOutLine.getM_Locator_ID(), inOutLine.getM_Product_ID(), asiID, null,
+															inOutLine.get_TrxName());
+
+		if (storageOnHand != null)
+		{
+			if (requiredQty.compareTo(storageOnHand.getQtyOnHand()) > 0)
+				return storageOnHand.getQtyOnHand();
+			else
+				return requiredQty;
+		}
+		return Env.ZERO;
+	} // getInOutLineQTY
+
+	/**
+	 * Get PO Costs in Currency of AcctSchema
+	 * 
+	 * @param  as Account Schema
+	 * @return    Unit PO Cost
+	 */
+	public BigDecimal getPOCost(MAcctSchema as, int inOutLineID, BigDecimal lineQty)
+	{
+		BigDecimal retValue = null;
+		//	Uses PO Date
+		String sql = "SELECT currencyConvert(ol.PriceActual, o.C_Currency_ID, ?, o.DateOrdered, o.C_ConversionType_ID, ?, ?) "
+				+ "FROM C_OrderLine ol"
+				+ " INNER JOIN M_InOutLine iol ON (iol.C_OrderLine_ID=ol.C_OrderLine_ID)"
+				+ " INNER JOIN C_Order o ON (o.C_Order_ID=ol.C_Order_ID) "
+				+ "WHERE iol.M_InOutLine_ID=?";
+		PreparedStatement pstmt = null;
+		ResultSet rs = null;
+		try
+		{
+			pstmt = DB.prepareStatement(sql, as.get_TrxName());
+			pstmt.setInt(1, as.getC_Currency_ID());
+			pstmt.setInt(2, as.getAD_Client_ID());
+			pstmt.setInt(3, as.getAD_Org_ID());
+			pstmt.setInt(4, inOutLineID);
+			rs = pstmt.executeQuery();
+			if (rs.next())
+			{
+				retValue = rs.getBigDecimal(1);
+				if (log.isLoggable(Level.FINE)) log.fine("POCost = " + retValue);
+			}
+			else
+				log.warning("Not found for M_InOutLine_ID=" + inOutLineID);
+		}
+		catch (Exception e)
+		{
+			log.log(Level.SEVERE, sql, e);
+		}
+		finally
+		{
+			DB.close(rs, pstmt);
+			pstmt = null; rs = null;
+		}
+		if (retValue != null)
+			retValue = retValue.multiply(lineQty);
+		return retValue;
+	}	//	getPOCost();
 }	//	MInOutLine
